+++ packages/google_mobile_ads/lib/src/ad_listeners.dart
@@ -210,7 +210,6 @@ class NativeAdListener extends AdWithViewListener {
     AdEventCallback? onAdImpression,
     OnPaidEventCallback? onPaidEvent,
     AdEventCallback? onAdClicked,
-    this.onNativeAdClicked,
   }) : super(
             onAdLoaded: onAdLoaded,
             onAdFailedToLoad: onAdFailedToLoad,
@@ -221,10 +220,6 @@ class NativeAdListener extends AdWithViewListener {
             onPaidEvent: onPaidEvent,
             onAdClicked: onAdClicked);
 
-  /// Called when a click is recorded for a [NativeAd].
-  /// @Deprecated Use [onAdClicked] instead
-  @Deprecated('Use [onAdClicked] instead')
-  final void Function(NativeAd ad)? onNativeAdClicked;
 }
 
 /// Callback events for for full screen ads, such as Rewarded and Interstitial.