+++ packages/google_mobile_ads/CHANGELOG.md
@@ -11,6 +11,8 @@
   [Admob](https://developers.google.com/admob/flutter/test-ads) and 
   [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads)
   documentation for up to date test ad units.
+* Deletes `NativeAdListener.onNativeAdClicked`. You should use `onAdClicked`
+  instead, which present on all ad listeners.
 
 ## 1.3.0
 * Adds support for programmatically opening the debug options menu using`MobileAds.openDebugMenu(String adUnitId)`